Переносит функции для рендеринга в отдельный модуль
