--- src/utils.js
@@ -1,30 +1,0 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforend`,
-};
-
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours());
-  const minutes = castTimeFormat(date.getMinutes());
-  return `${hours}:${minutes}`;
-};
-
-const createElement = (template) => {
-  const divElement = document.createElement(`div`);
-  divElement.innerHTML = template;
-  return divElement.firstChild;
-};
-
-const render = (container, element, place = RenderPosition.BEFOREEND) => {
-  if (place === RenderPosition.AFTERBEGIN) {
-    container.prepend(element);
-    return;
-  }
-  container.append(element);
-};
-
-export {RenderPosition, formatTime, createElement, render};
