+++ src/utils/common.js
@@ -1,0 +1,12 @@
+const castTimeFormat = (value) => {
+  return value < 10 ? `0${value}` : String(value);
+};
+
+const formatTime = (date) => {
+  const hours = castTimeFormat(date.getHours());
+  const minutes = castTimeFormat(date.getMinutes());
+  return `${hours}:${minutes}`;
+};
+
+
+export {formatTime};
